changed

+++ RaccoonEngineV2/src/raccoon/Portal.java
@@ -1,7 +1,7 @@
 package raccoon;
 
 public class Portal {
-	int x1, z1, x2, z2;
+	double x1, z1, x2, z2;
     int sector_a = -1;
     int sector_b = -1;
     String bottom_texture;
@@ -17,7 +17,7 @@ public class Portal {
     int top_tiled;
     boolean top_skip_texture;
     
-    public Portal(int x1, int z1, int x2, int z2, int sector_a, int sector_b, String bottom_texture, int bottom_brightness, int bottom_tiled, boolean bottom_skip_texture, String middle_texture, int middle_brightness, int middle_tiled, boolean middle_skip_texture, String top_texture, int top_brightness, int top_tiled, boolean top_skip_texture) {
+    public Portal(double x1, double z1, double x2, double z2, int sector_a, int sector_b, String bottom_texture, int bottom_brightness, int bottom_tiled, boolean bottom_skip_texture, String middle_texture, int middle_brightness, int middle_tiled, boolean middle_skip_texture, String top_texture, int top_brightness, int top_tiled, boolean top_skip_texture) {
     	this.x1 = x1;
     	this.z1 = z1;
     	this.x2 = x2;
