+++ RaccoonEngineV2/src/raccoon/Wall.java
@@ -2,14 +2,14 @@ package raccoon;
 
 public class Wall {
 
-	int x1, z1, x2, z2;
+	double x1, z1, x2, z2;
 	int sector_a = -1;
     String wall_texture;
     int wall_brightness;
     int wall_tiled;
     boolean skip_wall_texture;
     
-    public Wall(int x1, int z1, int x2, int z2, int sector_a, String wall_texture, int wall_brightness, int wall_tiled, boolean skip_wall_texture) {
+    public Wall(double x1, double z1, double x2, double z2, int sector_a, String wall_texture, int wall_brightness, int wall_tiled, boolean skip_wall_texture) {
     	this.x1 = x1;
     	this.z1 = z1;
     	this.x2 = x2;
